Fix rain data type and crash

+++ src/data_types.rs
@@ -40,7 +40,7 @@ pub(crate) struct OneDayForecast {
     pub(crate) time: i64,
     pub(crate) sunrise: i64,
     pub(crate) sunset: i64,
-    pub(crate) rain: i64,
+    pub(crate) rain: f64,
 }
 
 impl ProtoAdapter for OneDayForecast {
